Convert Android implementation to KMP-compatible
