--- composeApp/src/androidMain/kotlin/codes/chrishorner/personasns/theme.kt
@@ -1,12 +1,0 @@
-package codes.chrishorner.personasns
-
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-
-val OptimaNova = FontFamily(
-  Font(R.font.optima_nova_black, weight = FontWeight.Black)
-)
-
-val PersonaRed = Color(0xFFC41001)
