--- composeApp/src/androidMain/res/drawable/bg_splatter_background.xml
@@ -1,46 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="1440dp"
-    android:height="2560dp"
-    android:viewportWidth="1440"
-    android:viewportHeight="2560">
-  <path
-      android:pathData="M0,1272.3L0,1189.6L24.5,1173.8C27.2,1171.6 30.5,1168.1 31.3,1163.4C32.1,1158.5 29.5,1157.3 29.3,1152.1C29,1142 38.6,1139.2 37.8,1129.8C37.3,1124.4 33.9,1122.8 35.3,1118.8C36.7,1114.6 40.8,1115.2 43.2,1110.9C46,1105.9 42.4,1101.9 43.2,1094.5C43.6,1090.9 45.1,1085.9 50.3,1080.1L53.7,1081.5C52.8,1084.2 52.1,1087.7 53.7,1090.8C55.3,1093.9 57.8,1093.5 59.6,1096.2C63.6,1101.9 56.1,1109.3 59.6,1114.8C62.4,1119.3 67.3,1120.1 70.4,1125.3C72.3,1128.4 73.7,1129.5 72,1132.6C69.1,1138 63.5,1136.4 60.1,1142.5C58.9,1144.7 59.1,1145.9 57,1149.8C55.6,1152.4 54.2,1154.3 53.3,1155.4C46.7,1164.1 44.2,1165.7 40.3,1171.5C37.5,1175.7 35.5,1178.8 34.7,1183.4C33.8,1188.2 34.9,1191.4 36,1196.7C36.7,1200.1 39,1210.5 39,1220.7C39,1227.6 38.5,1233.7 33,1241.2C31,1243.9 28.4,1247.2 26.8,1249L20.1,1256.3L9.7,1264.7L0,1272.3Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-  <path
-      android:pathData="M69.7,1181.8C74.5,1178.2 80.1,1179 81.4,1179.2C86.3,1180 89.3,1182.9 89.9,1183.4C91,1184.5 95.4,1188.8 94.1,1193.2C93.1,1196.8 89,1198.3 86.9,1199.1C82.7,1200.7 79.1,1200 73.9,1199.1C69.2,1198.3 66.8,1197.8 65.4,1195.8C63.1,1192.5 65.1,1187.7 65.4,1187C66.6,1184.1 68.7,1182.5 69.6,1181.8L69.7,1181.8Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-  <path
-      android:pathData="M97.1,1105C99.4,1106.2 101.3,1106.5 103.9,1107C107.4,1107.6 113,1108.5 118.9,1106C121.3,1105 124.5,1103.7 126.1,1100.5C127.2,1098.2 128.2,1094.4 127.6,1089.7C125.6,1076.3 125.3,1077.9 123.5,1075.4C120.2,1070.7 114.4,1069.5 111.6,1068.9C109.5,1068.4 106.1,1067.7 102,1068.9C97.8,1070.1 95.2,1072.7 92.6,1075.4C90,1078 88.7,1079.3 87.7,1081.6C84.8,1088.4 88.5,1095.4 89.3,1096.9C90.1,1098.4 92.3,1102.5 97.1,1105Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-  <path
-      android:pathData="M979.9,83.9C980.1,83.5 981.3,80.5 979.9,78.2C977.5,74.2 970,76 969.5,76.1C967.8,76.5 964.8,77.2 963.1,80C961.6,82.4 961.2,85.9 963.1,88.6C963.3,88.9 964.6,90.6 966.8,91.2C970.1,92.1 973,90.1 975.6,88.4C977.5,87.1 979.1,86 979.9,84L979.9,83.9Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-  <path
-      android:pathData="M1045.1,66.5C1047.4,63.9 1051.1,63.7 1053.3,63.5C1056.1,63.3 1058.4,63.9 1059.8,64.4C1062.8,65.4 1065.2,65.4 1066.8,65.3C1068.7,65.1 1069.2,64.8 1075.2,63.6C1077.3,63.2 1078.8,62.9 1080.3,63.6C1080.9,63.9 1081,64 1082.9,65.3C1085.7,67.2 1085.8,67 1086.4,67.7C1088.3,69.9 1087.6,73.1 1087.5,73.3C1087.2,74.7 1086.4,75.7 1086,76.3C1082.9,80.2 1080.4,83.3 1075.9,85.1C1072.6,86.5 1070,86.4 1065.2,86.3C1063.6,86.3 1054.6,86 1046.1,81.4C1043.9,80.2 1041.7,78.7 1041.5,76.6C1041.5,75.8 1041.7,75 1042.4,73.2C1044,69 1044.8,66.9 1045.1,66.6L1045.1,66.5Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-  <path
-      android:pathData="M1173,64.8C1174.5,63.4 1176.6,63.5 1178.5,63.6C1181.4,63.7 1181.4,64.6 1184.1,64.8C1187.1,65 1187.5,64 1190.7,64.2C1192.4,64.3 1194,64.5 1195.4,65.6C1197.5,67.3 1197.7,70.2 1197.7,71.8C1197.7,73.2 1197.9,75 1196.9,76.8C1195.4,79.5 1192.4,80.3 1189.5,81.1C1186.3,82 1182.4,83.1 1178.4,81.1C1175,79.5 1173.5,76.6 1173,75.8C1171.3,72.7 1171.4,69.8 1171.5,69.2C1171.6,67.6 1171.7,65.9 1173,64.6L1173,64.8Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-  <path
-      android:pathData="M1243.9,1418C1243.9,1417.6 1243.7,1413.4 1246.4,1411.3C1247.9,1410.1 1249,1410.6 1252.2,1409.5C1255.3,1408.4 1255.6,1407.5 1258.2,1406.4C1262.1,1404.7 1265.7,1404.8 1268.3,1404.8C1272.8,1404.9 1276.1,1405 1278.4,1407.5C1279.8,1409 1280.9,1411.6 1280.4,1414C1279.9,1416.3 1278.2,1417.2 1273.9,1420.7C1267.7,1425.8 1268.9,1425.8 1266.7,1427C1262.4,1429.4 1255.6,1430.3 1250.1,1427C1249.2,1426.5 1244.1,1423.4 1243.8,1418L1243.9,1418Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-  <path
-      android:pathData="M1162.1,1442.6C1162.7,1442.1 1165.7,1440.1 1169.3,1440.8C1170,1440.9 1173.9,1441.8 1174.5,1444.6C1175,1446.9 1172.9,1448.9 1172.5,1449.3C1169.7,1452 1165.2,1452 1162.2,1450.4C1161.3,1449.9 1159.6,1449 1159.3,1447.3C1159,1445.6 1160.2,1443.8 1162.2,1442.6L1162.1,1442.6Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-  <path
-      android:pathData="M1384.6,1795.3C1385.7,1792.7 1387.9,1791.6 1391.8,1789.7C1395.1,1788 1397.1,1787 1400.2,1786.8C1404.4,1786.5 1407.4,1787.8 1409.7,1788.7C1412.9,1790 1415.4,1791.1 1415.9,1793.4C1416.3,1795.1 1415.3,1796.5 1413.6,1799C1412,1801.4 1410.6,1803.4 1408.4,1804.6C1407,1805.4 1405.5,1805.7 1396.9,1804.6C1386.9,1803.3 1385.4,1802.5 1384.7,1801.1C1383.3,1798.7 1384.5,1795.7 1384.7,1795.3L1384.6,1795.3Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-  <path
-      android:pathData="M0,1280L0,1366C4,1358.9 6.1,1352.5 7.3,1347.9C8.3,1344 9.4,1339.4 9.5,1333.2C9.5,1326.5 8.3,1325.1 9.5,1321C11.3,1315 15.1,1314.5 15.1,1310.2C15.1,1306.3 11.9,1306.4 8.7,1299.9C5.9,1294.1 7.2,1291.3 3.8,1285.2C2.4,1282.8 1,1281 -0.1,1279.8L0,1280Z"
-      android:fillColor="#ffffff"
-      android:fillType="nonZero"/>
-</vector>
