+++ composeApp/src/commonMain/composeResources/drawable/next.xml
@@ -1,0 +1,25 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="130dp"
+    android:height="57dp"
+    android:viewportWidth="130"
+    android:viewportHeight="57">
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M47.24,3.82L60.8,4.69L64.65,10.83L65.92,4.01L77.18,7.1L78.38,12.84L87.26,9.16L91.69,15.31L101.05,15.47L101.93,8.47L114,4.01L115.68,12.37L119.69,12.49L122.13,20.81L119.08,23.79L119.38,29.48L126.33,34.4L117.97,47.66L109.87,42.77L105.98,44.83L100.63,44.72L98.64,43.38L95.89,44.98L85.81,45.22L84.4,41.66L80.18,43.05L78.37,48.19L79.97,51.67L65.42,52.2L63.17,47.24L60.61,51.21L46.67,51.51L47.24,3.82Z" />
+    <path
+        android:pathData="M63.07,2.22C63.34,1.79 63.7,1.43 64.13,1.16C64.93,0.66 65.9,0.52 66.81,0.77L78.07,3.86C79.29,4.19 80.22,5.18 80.47,6.41C80.47,6.41 80.84,8.18 80.84,8.18L85.98,6.06C87.42,5.46 89.08,5.93 89.99,7.19C89.99,7.19 93.44,11.97 93.44,11.97C93.44,11.97 98.09,12.05 98.09,12.05L98.59,8.06C98.75,6.81 99.58,5.76 100.76,5.32L112.83,0.85C113.75,0.51 114.78,0.59 115.64,1.07C116.49,1.55 117.1,2.38 117.29,3.35L118.45,9.09L119.78,9.12C121.24,9.17 122.5,10.14 122.91,11.54L125.36,19.86C125.71,21.06 125.37,22.35 124.48,23.22L122.52,25.13L122.65,27.67L128.28,31.66C129.73,32.69 130.13,34.69 129.18,36.2L120.81,49.45C119.84,51 117.8,51.49 116.23,50.54L109.76,46.63L107.55,47.8C107.04,48.07 106.48,48.21 105.91,48.19L100.56,48.08C99.91,48.06 99.29,47.86 98.75,47.51L98.52,47.35L97.59,47.89C97.1,48.18 96.54,48.33 95.97,48.35L85.89,48.58C84.48,48.61 83.21,47.76 82.69,46.46L82.61,46.27L81.99,48.02L83.02,50.26C83.49,51.28 83.42,52.47 82.84,53.43C82.25,54.38 81.22,54.99 80.09,55.03L65.54,55.56C64.32,55.61 63.18,54.98 62.55,53.95C62.01,54.34 61.36,54.56 60.68,54.57L46.74,54.88C45.83,54.9 44.95,54.54 44.3,53.91C43.65,53.26 43.29,52.38 43.31,51.47L43.88,3.78C43.89,2.85 44.28,1.98 44.96,1.35C45.63,0.72 46.54,0.4 47.46,0.46L61.02,1.34C61.79,1.39 62.51,1.7 63.07,2.22ZM47.24,3.82L46.67,51.51L60.61,51.21L63.17,47.24L65.42,52.2L79.97,51.67L78.37,48.19L80.18,43.05L84.4,41.66L85.81,45.22L95.89,44.98L98.64,43.38L100.63,44.72L105.98,44.83L109.87,42.77L117.97,47.66L126.33,34.4L119.38,29.48L119.08,23.79L122.13,20.81L119.69,12.49L115.68,12.37L114,4.01L101.93,8.47L101.05,15.47L91.69,15.31L87.26,9.16L78.38,12.84L77.18,7.1L65.92,4.01L64.65,10.83L60.8,4.69L47.24,3.82Z"
+        android:fillColor="#ffffff" />
+    <path
+        android:pathData="M53.04,11.57C53.07,10.22 55.26,9.14 56.87,9.14C58.49,9.14 60.49,9.25 61.11,11.42C61.72,13.59 68.5,34.73 68.5,34.73L67.84,10.76C67.83,9.42 69.95,8.78 71.11,8.78C72.27,8.78 74.5,9.51 74.5,11.8C74.5,14.1 74.79,22.54 74.79,22.54C75.98,19.06 79.39,15.19 82.66,15.19C85.92,15.19 88.94,16.86 90.1,19C92.48,19 95.81,19.36 95.81,19.36L98.01,23.43L100.76,19.4L103.66,19.27L106.26,17.84L106.29,11.24C106.3,9.74 108.37,8.76 109.53,8.76C110.69,8.76 112.16,9.47 112.16,10.96C112.16,12.46 112.74,15.55 112.74,15.55C112.74,15.55 114.2,16.08 114.94,15.97C115.67,15.85 116.22,17.6 116.22,18.94C116.22,20.29 115.76,22 113.74,22.73C113.74,24.22 113.86,28.88 113.96,30.46C114.06,32.03 115.12,33.51 116.8,33.51C118.48,33.51 118.82,33.79 118.82,34.8C118.82,35.8 119.52,41.03 114.94,41.03C110.35,41.03 108.43,35.5 108.43,32.42C108.43,29.33 108.98,25.36 106.72,24.96C106.24,24.88 105.76,23.47 105.4,23.62C104.06,24.14 103.61,26.63 102.41,28.01C102.07,28.71 102.13,29.2 102.13,29.2C102.13,29.2 106.53,36.68 108.46,38.99C108.46,40.14 107.85,40.78 104.46,40.78C101.06,40.78 101.53,40.16 100.36,38.99C99.2,37.83 98.41,35.25 98.41,35.25C97.25,36.94 96.4,38.05 95.97,38.99C95.55,39.93 94.99,41.15 93.12,41.15C91.26,41.15 88.34,41.08 87.9,38.41C87.04,40.13 84.44,40.2 82.46,40.2C80.47,40.2 77.45,40.23 75.16,35.16C75.16,37.18 75,42.98 75,44.48C75,45.98 74.27,47.6 71.61,47.6C68.96,47.6 69.2,47.36 67.95,47.29C64.98,47.11 61.61,28.79 59.07,21.42C59.16,25.99 59.26,42.5 59.26,43.79C59.25,45.07 59.64,47.69 57.13,47.69C54.63,47.69 53.37,47.47 53.37,44.63C53.37,41.79 53,13.82 53.04,11.57ZM94.49,29.31C94.25,28.35 92.44,25.51 91.9,24.98C91.75,26.3 91.14,28.82 88.4,28.82C85.67,28.82 80.94,29.86 79.28,29.15C79.72,30.3 80.81,33.84 82.83,33.84C84.85,33.84 85.35,31.81 87.07,31.81C89.44,31.81 90.98,33.55 89.85,35.77C90.9,34.33 93.61,31.11 94.49,29.31ZM82.87,20.97C81.86,20.97 80.83,22.46 80.83,23.2C80.83,23.93 82.15,24.29 82.95,24.29C83.76,24.29 84.91,23.8 84.91,23.02C84.91,22.24 84.11,20.97 82.87,20.97Z"
+        android:fillColor="#ffffff" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M24.72,27.48m-21.36,0a21.36,21.36 93.07,1 1,42.71 0a21.36,21.36 93.07,1 1,-42.71 0" />
+    <path
+        android:pathData="M24.72,2.75C38.36,2.75 49.44,13.83 49.44,27.48C49.44,41.12 38.36,52.19 24.72,52.19C11.08,52.19 -0,41.12 -0,27.48C-0,13.83 11.08,2.75 24.72,2.75ZM24.72,6.12C12.93,6.12 3.36,15.69 3.36,27.48C3.36,39.26 12.93,48.83 24.72,48.83C36.51,48.83 46.08,39.26 46.08,27.48C46.08,15.69 36.51,6.12 24.72,6.12Z"
+        android:fillColor="#ffffff" />
+    <path
+        android:pathData="M21.02,38.99L16.91,36.94L26.37,27.48L16.91,18.02L21.02,15.96L32.53,27.48L21.02,38.99Z"
+        android:fillColor="#ffffff"
+        android:fillType="nonZero" />
+</vector>
