+++ composeApp/src/commonMain/composeResources/drawable/ryuji.png
