+++ composeApp/src/commonMain/composeResources/drawable/yusuke.png
