+++ composeApp/src/commonMain/kotlin/codes/chrishorner/personasns/animations.kt
@@ -10,9 +10,9 @@ const val AnimationDurationScale = 1
 
 // Temp work around for https://issuetracker.google.com/issues/354405919
 val BetterEaseOutBack: Easing = Easing { fraction ->
-  try {
-    EaseOutBack.transform(fraction)
-  } catch (e: IllegalArgumentException) {
-    1f
-  }
+    try {
+        EaseOutBack.transform(fraction)
+    } catch (e: IllegalArgumentException) {
+        1f
+    }
 }
