+++ composeApp/src/commonMain/kotlin/codes/chrishorner/personasns/random.kt
@@ -6,13 +6,13 @@ import androidx.compose.ui.unit.dp
 import kotlin.random.Random
 
 fun randomBetween(start: Float, end: Float): Float {
-  return start + Random.nextFloat() * (end - start)
+    return start + Random.nextFloat() * (end - start)
 }
 
 fun randomBetween(start: Dp, end: Dp): Dp {
-  return randomBetween(start.value, end.value).dp
+    return randomBetween(start.value, end.value).dp
 }
 
 fun Density.randomPxBetween(start: Dp, end: Dp): Float {
-  return randomBetween(start, end).toPx()
+    return randomBetween(start, end).toPx()
 }
